requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0
packaging
sentencepiece==0.1.99
accelerate==0.21.0
peft
bitsandbytes
pydantic
shortuuid
requests
ninja
tokenizers==0.15.1
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
markdown2==2.4.10
numpy==1.23.5
openai==1.3.7
opencv_python==4.9.0.80
scikit-learn==1.2.2
gradio==4.16.0
gradio_client==0.8.1
Pillow==9.4.0
ray==2.6.1
tqdm==4.64.1
transformers==4.37.2
uvicorn
fastapi
scipy==1.11.2
deepspeed==0.12.6